+++ deploy/rules/RDiamondCubePostSetup.js
@@ -2,7 +2,7 @@ const { get, getAlias, registerRule } = require('../utils/deploy.js');
 
 module.exports = async () => {
   const feeRuleRegistry = await get('FeeRuleRegistry');
-  const rule = await getAlias('RCubeNFT', 'RDiamondCube');
+  const rule = await getAlias('RStarNFTV4', 'RDiamondCube');
   await registerRule(feeRuleRegistry, rule);
 };